requirements.txt
pandas>=2.0.3
psycopg2_binary>=2.9.3
